mvoicefinal/pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-ai-voicelive>=1.1.0",
    "azure-core>=1.35.0",
    "azure-identity>=1.24.0",
    "dotenv>=0.9.9",
    "pyaudio>=0.2.14",
]
